config: add .env variables
