+++ .npmignore
@@ -3,3 +3,4 @@ node_modules
 build.js
 dist/*.html
 dist/*.min.js
+.env
