+++ root.d.ts
@@ -1,0 +1,3 @@
+interface ImportMeta {
+  readonly env: Record<string, string>;
+}
