+++ tsconfig.json
@@ -11,7 +11,8 @@
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true,
     "removeComments": true,
-    "moduleResolution": "node"
+    "moduleResolution": "node",
+    "types": ["./root.d.ts"]
   },
   "include": ["src"]
 }
